Add tests, and a small refactor
Mostly, we're pulling the GORM Type methods out of the individual
types, moving them to a common mixin type instead.

Testing revealed some minor bugs, as well as some extensive limitations
of the ASN1 and XML serializers, so the bugs were fixed, and
support for ASN1/XML removed for the time being.